database: Don't connect to replica set unless explicitly told to
The semantics of the MONGODB_REPLICASET environment variables used
to be weird; connect to "rs0" if MONGODB_REPLICASET contained any
non-empty string, otherwise don't connect to a replica set. Change
this to default to not using a replica set but to use the string in
MONGODB_REPLICASET as the replica set.