storage: Use event id as document id when inserting to MongoDB
When storing an event in a MongoDB collection don't let MongoDB
allocate a random document id. Instead use the event id as the
document id and catch any DuplicateKeyError errors that might pop
up. Unless you have events with duplicate ids sent on the bus
such errors are harmless and typically indicate that a message
was delivered twice.

We also change insert_to_db() to return a bool indicate whether
a document was actually inserted. If nothing else this is useful
in the tests.