Fix broken tests (#35)
Two recent commits, b1ea2a5 and d248c4c each broke the tests in
different ways. Most failed tests stemmed from the latter commit
and were trivial to address.

The first commit refactored storage.py and moved some functions
to database.py, so we move the test file accordingly. The API of
insert_to_db() was also changed to return True if the document was
ignored because it was already present in the database. That made
sense but broke the test which relied on the return value to
determine whether the document was inserted. It was rewritten to
count the documents in the collection to make that determination.
This prompted the addition of a pytest fixture, mock_mongo, that
injects a fresh MongoDB database into the test.