storage.py: Support RABBITMQ_DURABLE_QUEUE env variable
There are very few reasons why one wouldn't want to use a durable
queue when consuming events and storing in a database and previously
we haven't supported that. Since flipping the durable queue parameter
is a breaking change we make it opt-in via a new RABBITMQ_DURABLE_QUEUE
environment variable. If we ever flip the program's major version we
can change its default to true.