storage.py: Trap SIGTERM and shut down cleanly
eiffellib's RabbitMQ threads prevent the program from shutting down
when it receives a SIGTERM signal. This is annoying in general, and
when run in a process manager (like Docker) that expects the process
to shut down after receiving SIGTERM the end result is probably a slow
shutdown when the process manager patiently waits for a response to
SIGTERM and, when there's no reaction, issues a SIGKILL signal. As a
bonus the busy wait loop in main() is gone.