Add dockerfiles for graphql-api and storage
The added dockerfiles use multi-stage image builds to first build a
wheel of the code and then installing the wheel in the second stage
which uses a smaller base image. While it's the same Python code
being installed in both cases, graphql-api and storage use different
entrypoints and only the former is exposed to the network.